LayoutTests/http/wpt/webauthn/public-key-credential-create-failure.https-expected.txt

CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.create' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.create' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.create' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.create' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.get' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.create' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.get' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.create' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.get' within user activated events.

PASS PublicKeyCredential's [[create]] with timeout 
PASS PublicKeyCredential's [[create]] with a mismatched RP ID 
PASS PublicKeyCredential's [[create]] with an empty pubKeyCredParams 
PASS PublicKeyCredential's [[create]] with two consecutive requests 
PASS PublicKeyCredential's [[create]] with two consecutive requests (2) 
PASS PublicKeyCredential's [[create]] with new requests in a new page 
PASS PublicKeyCredential's [[create]] with new requests in a new page (2) 
